update content worker to new system
