+++ web/modules/cm-lang-ftml/package.json
@@ -16,11 +16,11 @@
     "@types/prismjs": "^1.26.0",
     "@wikijump/cm-espells": "workspace:*",
     "@wikijump/codemirror": "workspace:*",
+    "@wikijump/comlink": "workspace:*",
     "@wikijump/components": "workspace:*",
     "@wikijump/fluent": "workspace:*",
     "@wikijump/ftml-wasm-worker": "workspace:*",
     "@wikijump/prism": "workspace:*",
-    "@wikijump/threads-worker-module": "workspace:*",
     "@wikijump/util": "workspace:*",
     "cm-tarnation": "^0.1.1",
     "svelte": "^3.46.4"
